backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
faiss-cpu
numpy
openai        
tiktoken
PyPDF2
boto3
pytesseract
pdf2image
Pillow